Fix rotate-detection rules to be less sensitive when unit is flat.
The rotation detection rules were too simple, and too general, and
caused the system to flicker madly if you weren't holding it upright
at more than a 40° angle.  This was mostly due to my misreading the
original algorithm.  This instance currently matches Poel's algorithm
and remembers to quit once a matching geometry has been found.  It's
much more stable and pleasant to use.